backend/requirements.txt
fastapi>=0.111.0
pydantic>=2.7.0
uvicorn>=0.29.0
python-multipart>=0.0.9
paho-mqtt==1.6.1
influxdb-client==1.36.1
SQLAlchemy>=2.0.20
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
prometheus_client==0.17.1
pytest==7.4.2
